+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-- Add ServerConnectionEvents.
-- Add capabilities handshake system.
-- Add logical server lifecycle events (#38).
+- Fix network capability crash when connecting to a modded server (#41).